// apps/web/types/canva-types.ts
export type ActiveElement = {
  name: string;
  value: string;
  Icon: LucideIcon | null;
} | null;
// ...
export type NavbarProps = {
  activeElement: ActiveElement;
  imageInputRef: React.MutableRefObject<HTMLInputElement | null>;
  handleImageUpload: (e: React.ChangeEvent<HTMLInputElement>) => void;
  handleActiveElement: (element: ActiveElement) => void;
};